Change custom elements definitionFn API
definitionFn has two arguments: User and Properties. User replaces
previous `element` argument which was creating additional boilerplate
(User was always defined in the same way in custom elements). This is
a breaking change. It also prepares for the upcoming HTMLRenderer,
which in the context of server-side rendering, could be given as the
"User" in a custom element definitionFn.