Fix missing rootElem$ emissions on Widget #87
The real problem was: when an existing custom element was recreated
in the virtual DOM using the custom element's Widget constructor,
the new Widget instance did not have a reference to the rootElem$
of the custom element's DOMUser. This rootElem$ is needed for the
parent DOMUser of the custom element which needs to wait for all
children rootElem$ before emitting its own rootElem$. The 2nd Widget
instance did not have a reference to its rootElem$ because that
was created by and retained at the 1st Widget instance which calls
init() since it is the first. init() had the construction of the
custom element's rootElem$. Now, the Widget constructor has the
construction of the custom element's rootElem$, which replicates
the original rootElem$ (attached to the Widget's DOM element). We
attach the DOMUser to the widget's DOM element just like props are
attached to the DOM element. The DOMUser needs to make sure diff&patch
doesn't wipe away both props and DOMUser from the widget DOM element.

Conflicts:
	dist/cycle.js
	dist/cycle.min.js
	src/dom-user.js