fix(EventDelegator): avoid bubbling for dead element event
While bubbling, some destination event handler may trigger cascading
changes in the application so that the element that owned the event
handler ends up being synchronously removed from the DOM. Then,
immediately/synchronously after that, the event handler of the parent
element of the zombie element would be triggered simply based on the
selector of the parent being the same of the child (think a simple
one-class selector for an application that has nested folders, the
parent component is the same type of the child component), and this
would also trigger cascading changes in the application, this time for
wrong information, because the event should not propagate anymore.

The fix simply checks if the element owning the event is actually
existing in the DOM, before bubbling.