fix(DOMSource): make DOMSource an interface
Make DOMSource be an interface, not a class. The previous class
DOMSource is now the class MainDOMSource. HTMLSource is a DOMSource, and
so is MockedDOMSource a DOMSource. DOMSource.elements getter was
converted to DOMSource.elements() thunk function, for consistency with
DOMSource.events(). New file root/xstream-typings.d.ts provides custom
DOMSource typings for xstream. Other stream libraries should follow
later on.