fix(dom): use isolation scope as snabbdom key
This commit fixes a bug when building a list of isolated components. Due to how the isolate module
works, if a component element was removed, then if the component element doesnt have a key property,
then isolate module will understand that the removed element was not actually removed, but its
contents were swapped with the component below it. This would cause both the removed and the below
elements to have their isolation scopes wiped out in the isolate module. Then, the remaining element
could not be clicked. This bug was reproduced in manual-tests/advanced-list and a new test case was
added too.