refactor(history): fix/undo broken code from 2d3bb3c
If we had released history driver with commit 2d3bb3c, that would be a bug since the tests were not

passing. Because we did not release, this commit undoes the bugs that 2d3bb3c introduced, so +1 -1

means there is no new release that needs to be done, and in total, we just refactored code.