fix(dom): change EventDelegator bubble to search topElement
Searching the body will not find the element inside a DocumentFragment or shadowRoot, so the

function ends up exiting early and the event is ignored. Added a helper function to tests to

determine whether event bubbling or capture will work inside a DocumentFragment. Added two tests to

check bubbling and capture.

ISSUES CLOSED: #412