fix(dom): fix isolation bug of child when parent is re-added
This bug was not happening before when we had transposition on because transposition would

eventually clone all vnodes. The key to fix this bug was IsolateModule.scopesBeingUpdated, so that

we avoid cleaning up the event delegators in IsolateModule.cleanupVNode() if we know that the

element is being replaced instead of actually removed. This is because in snabbdom, when an element

is replaced by another, the creation of the new element happens BEFORE the destroy of the old

element, so we dont want the destroyal of the old to wipe out good information concerning the new

element (such as updating its top element in the EventDelegator).