fix(jsonp): use wildcard version for cycle/run
Use "cycle/run": "*" as the dependency, because if the user already has
cycle/run in his package.json, we want to use exactly the same version
of run, because we need to have the same adapt module. Otherwise we will
have two different run versions installed, which means different state,
and a bug.