test(history): update saucie to v3.1.0, refactor tests dir
Update Saucie to the latest version. Also reconfigure testem to spawn a new page for three different
specs, so that history state is isolated between these 3 cases, and doesnt mess up with each other.
Also, we turned off CI tests because testem tests running in the browser lose connection (socket.io
perhaps) to testem once the history is modified. This is an unsolved problem that we need to fix
later.