fix(run): fix race condition for drivers that subscribe late
The code fixed in src was necessary to make the newly added test pass.
Sink proxies should be MemoryStreams in order to fix race conditions
for drivers that subscribe to sink proxies "later".
Recall that there are two steps:

1. Setup (sink proxies -> drivers -> sources -> main -> sink)
2. Execution (also known as replication: sink proxies imitate
sinks)

If a driver does not synchronously/immediately subscribe to
the sink proxy in step (1), but instead does that later, if
step (2) feeds a value from the sink to the sink proxy, then
when the driver subscribes to the sink proxy, it should receive
that value. This is why we need MemoryStreams, not just Streams.
Note: Cycle DOM driver is an example of such case, since it waits
for 'readystatechange'.