fix(run): type check keyof drivers and main with TypeScript 2.2
Update to TypeScript 2.2 and use the keyof feature to construct the expected type of Drivers object
from Sources and Sinks fromain function.

BREAKING CHANGE:
If you are using JavaScript, literally nothing has changed, and this is not a breaking release.
However, if you are using TypeScript, this version may catch errors and typos that previous Cycle
Run versions didnt. Also, this version expects that if Sinks of main() are typed, they must be a
type alias, not an interface. Usually you should leave the Sinks object implicitly typed, but we
also support type aliases if you want to explicitly type them.

ISSUES CLOSED: 538