Fix display of simultaneous events in assert errors
When parsing diagrams, simultaneous events emit at the left parenthesis,
and each character represents the passing of time. This is to ensure
that all diagrams line up vertically with a monospace font.

However, there was an issue where the logic that turns a test failure
into diagram strings was treating them as if they only took one frame.

Consider this test, which will fail:

    const Time = mockTimeSource();

    const A         = Time.diagram('--a-----a---');
    const B         = Time.diagram('--b-------b-');
    const expected$ = Time.diagram('--(ab)--a-c-');

    const actual$ = xs.merge(A, B);

    Time.assertEqual(actual$, expected$);

    Time.run(done);

Before this change, this test will fail with the following message:

    Expected

    --(ab)-----a-c

    Got

    --(ab)-----a-b

This does not match the diagram, which shows two blank spaces following
the simultaneous event. Now the correct message will be shown:

    Expected

    --(ab)--a-c

    Got

    --(ab)--a-b

This is consistent with the format of the string given to .diagram(),
and should make testing with simultaneous events more intuitive.

Thanks to @rjkip for reporting this issue :)