feat(isolate): allow a scopes-per-channel object as second arg
isolate(component, scopes) now allows scopes to be an object as a map
from key to scope, allowing giving a different scope for each isolation
channel. For instance, you can give the scope 'foo' to the DOM
source/sink and the scope 'bar' to the HTTP source/sink. This
scopes-per-channel also accepts the wildcard scope under the key '*',
which will work as the default isolation scope for channels not
specified in the scopes-per-channel object. If the wildcard is not
present, an arbitrary scope string will be generated.

ISSUES CLOSED: 526