feat(dom): allow choosing no isolation or sibling isolation
Allow the scope ':root' to mean no isolation, as an argument for isolateSource/isolateSink. Also
allow a className or id selector as the isolation scope to mean sibling-sibling isolation (which
means the parent can freely access the tree of its children as well).

BREAKING CHANGE:
Rare breaking change that will likely NOT affect you: we are giving special meaning to some scope
strings given to isolation. Before, every string given as isolation scope would mean total
isolation. Now, the string ':root' means no isolation and strings starting with '.' or '#' mean
sibling-sibling isolation. This is a technically a breaking change in case you happened to use
strings like that as isolation scopes (which is unlikely). But otherwise, this is safe to update
without migration.

ISSUES CLOSED: 526