feat(dom): allow predicate function or object as preventDefault
When using DOMSource.events(), you used to be able to pass a boolean as
option in the second argument, e.g. `events('click', {preventDefault:
true})`. Now, preventDefault option can be a function (taking the event
as input and returning a boolean) or an object describing the expected
properties on the event object.

ISSUES CLOSED: #611