feat(isolate): support null scope to disable isolation
When providing a null value as the isolation scope for some channel, that channel will not be
isolated. For instance, if you pass {DOM: null, HTTP: 'foo'} as the scopes to isolate(), then only
HTTP will be isolated, while DOM source and sink will stay like it was before. The wildcard '*'
scope can also be null to disable isolation for unspecified channels.

BREAKING CHANGE:
This is a breaking change only in case you utilized null scopes to perform isolation. Previously
null scope would enable isolation using null as the isolation name, and now null scope will just
disable isolation. It would be quite a corner case to rely on null scopes, so this will breaking
change likely not affect your project, and this new version of isolate is rather safe to upgrade.