fix(time): fix periodic dropping behind in background tab
Previously, if you compared both a periodic stream from @cycle/time, and
one from xstream, if you leave them running in a background tab for a
while, you will notice the @cycle/time periodic dropping behind the
xstream periodic.

This is because in @cycle/time, every time a periodic event occurs, the
next one is scheduled the given period ahead in time.

This fails to account for events occurring slower than the given
period, which can easily happen in a background tab. It is noticeable
even with a period as large as 1 second.

The fix is to schedule the next event for when it should be based
entirely on the period, unrelated to the time the current event fired.

With the fix in place, @cycle/time periodic actually runs faster than
xstream periodic in background tabs. If you call `performance.now()`,
you can see that the @cycle/time periodic is correct, and the xstream
periodic is behind.

In my testing, over the course of 10 minutes xstream fell ~30 seconds
behind real time. This is around a 7% innacuracy. This is because
xstream suffers from the same bug that existed in this library previous
to this change.