fix(dom): revert patchEvent optimization, to support IE10+
We had done a memory optimization in commit 0153cea891ec62327cf3a638aa0283e3bb0de156 to avoid
unnecessary function allocations, but it seems that we cant mutate event properties in IE10/IE11, so
in those browsers we experienced an infinite call stack on stopPropagation(). This commit reverts
that optimization.