fix(dom): wait for snabbdom's post hook to clean vnodes (#667)
Initially, vnodes were cleaned (EventListener destroyed) during the snabbdom's \`remove\` or \`destroy\`
hook. Thus, if the vnode is, in fact, used latter on in the DOM, those listeners are now gone.
This
patch only removes those nodes at the end of snabbdom's process. Thus if the nodes are needed latter
on, they won't be destroyed

ISSUES CLOSED: #667