fix(run): schedule sink emissions as microtasks
Every sink stream emission from the main() is scheduled on the microtask available (MutationObserver
or process.nextTick or setTimeout) in order to give enough time for subscriptions to happen in
main() and avoid race conditions such as issue #592.

BREAKING CHANGE:
This means Cycle run() function is no longer synchronous, so if for some reason you were depending
on the synchronicity of run(), you will lose this guarantee. This is breaking change, but rarely
should this be experienced. With the new microtask scheduling, sinks are emitted to drivers as soon
as the current event loop scripts are completed, so there will not be noticeable delays in the user
interface (like setTimeout would incur).

ISSUES CLOSED: #592