Simplify error handling API of cycle/http
The API of cycle/http conflates HTTP responses with error status and
xstream errors. Makes users of cycle/http use two different API
mechanisms from xstream to work with cycle/http. Normal xstream events
for HTTP responses with status codes in the 200 range and xstream errors
for HTTP responses outside of the 200 range. Timeouts etc use xstream
errors too.

The solution I pursued was to make all responses from Superagent sent as
normal xstream events. So now everything that comes out of cycle/http is
a normal xstream event and can be filtered in basic xstream
ways. replaceError() never needs to be used and no considerations for if
streams have been replaced needs to be made. Diamond paterns can still
be used if desired. If the user of the library wants xstream errors they
can do that too.

This makes the API of cycle/http better adhere to the principle of least
supprise. It also simplifis it since there is only one xstream API
mechanism that needs to be used instead of two.