fix(dom): remove "change" from eventTypesThatDontBubble (#690)
BREAKING CHANGE:
It's probable your app will not break when updating cycle/dom. However,
if you are using `domSource.events('change')` and unknowingly relying on
buggy behavior of this library, your app might behave differently. So
upgrade carefully if you have that use case, otherwise this version is
very safe to upgrade to, since this bug fix is so tiny.