fix(dom): use MutationObserver to avoid mutation loop
When vnode$ depends on element$ and vice-versa, a synchronous infinite loop might occur, which is
issue #699. This commit solves that by using MutationObserver, used to indicate when to emit a new
element on element$.

BREAKING CHANGE:
Internet Explorer 10 is no longer officially supported, but it can still
be used with cycle/dom under some circumstances. You should use a
polyfill for MutationObserver and make sure you are not rendering the
application in a DocumentFragment as the container node. Only under
those conditions will cycle/dom should work correctly in IE10.

ISSUES CLOSED: #699