fix(time): fix throttle immediately completing with sync producer
v0.12.0 introduced a regression where applying throttle to synchronous
source (like xs.of or xs.from) would cause the resulting stream to
immediately complete, before firing an event.

This was not a documented behaviour, but @cycle/time consistently
processes next events prior to completion events if they occur in the
same frame.

I have added a unit test to prevent any regressions to this behaviour in
the future.

Thanks to @willheslam for reporting this issue.

This fixes #770.