Do Not Duplicate Root Elements w/o ID
Do not duplicate the root element when not using ID property. For example, without this PR, when simply targeting the HEAD tag it gets duplicated because there is no associated ID. This PR simply gives the `selId` variable a default value of empty string, which correlate it with the default value of empty string given to the ID property of the `rootElement` object by `document.querySelector`.