fix(isolate): adapt sinks so xstream semantics can be applied
Rxjs 6 uses a piping semantic which removes the existence of the map prototype from the base
Observable class, this change ensures that each sink stream is adapted to xstream before
isolation
operations are applied.

BREAKING CHANGE:
This update changes the contract such that xstream and @cycle/run are now dependencies.

ISSUES CLOSED: #826