Fix improper import statement
'raf' and 'performance-now' modules export functions, and 'import *' produces a namespace object, which is always a plain object. This error is hidden in cjs compile targets, because require() calls return with the exact module.exports instance. But in ES6 targets, compilers throw an error saying that 'requestAnimationFrame is not a function'