fix(rxjs-run): use peerDependencies to avoid dependency shadowing
At some point in the past the peerDependencies usage was dropped in
favor of dependencies.  This
introduced a major bug causing the rxjs dep from
rxjs-run to shadow the rxjs version used in the
app which causes problem since
testing whether something is an Observable is done
via equality
testing on the Observale prototype.

ISSUES CLOSED: #902