feat(run): use bigint instead of number for ids
As one id is generated for every event, using a number would lead to a
wrap-around of ids. This might lead to very subtle bugs for long running
applications. As we are only incrementing the ids by one and not do any
arithmetic on them, using bigint does not create any performance impact